fix: ensure input_ids dtype is int64 in prepare_for_decode
Non-greedy sampling (sampling_from_probs_torch) returns int32 token IDs,
which violates the input_ids int64 contract and crashes hash_topk kernel
(hash_topk.cuh:116 only accepts int64). Enforce dtype at the
output_ids -> input_ids assignment boundary.

+++ python/sglang/srt/managers/schedule_batch.py
@@ -2019,7 +2019,7 @@ class ScheduleBatch(ScheduleBatchDisaggregationDecodeMixin):
                 )
 
         # Update fields
-        self.input_ids = self.output_ids
+        self.input_ids = self.output_ids.to(torch.int64)
         self.output_ids = None
 
         if self.model_config.is_encoder_decoder:
